extractor/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
cachetools==2.0.0
cssselect==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
google-auth==1.0.0
google-auth-httplib2==0.0.2
google-cloud-bigquery==0.24.0
google-cloud-core==0.24.1
googleapis-common-protos==1.5.2
httplib2==0.10.3
idna==2.5
jieba3k==0.35.1
lxml==3.7.3
newspaper3k==0.1.9
nltk==3.2.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.13.0
requests-file==1.4.2
rsa==3.4.2
six==1.10.0
tldextract==2.0.2
